Refactored test suite so that it isn't dependent on the data module.
This is needed so that components can be developed that aren't tighly
coupled to our project because originally test required knowledge of our
hive data models.
